fix(ios): prevent server stopping
I see a lot of "BlobWriter server not running" errors via Sentry, but
I am unable to reproduce it while the device is attached to a debugger.

I suspect it has something to do with the backgrounding of the app. If
the error does not disappear due to this commit, it should be reverted.