Revert "fix(ios): prevent server stopping"
This reverts commit b751e3c295834c8a4d5adaf24c590a0850ba1c1c.

The webview was not able to communicate with the server whilst the app was
backgrounded, even if the server continued running. When 'fetch' was
called, it would reject with a TypeError ("Could not connect to the
server.").

I prefer the "BlobWriter server not running." error as it is more
informative, so I am reverting this change.