Foreground
Sadly, it appears that iOS does not allow a web server to operate
indefinitely in the background. This caused a bug where the server
was unresponsive the next time the app came to the foreground.

This commit reverts cb517d6 and adds a fix so that the app does not
crash when started in the background.

close #49
close #53